Refactor k-headline
